fix(gateway): preserve global model authority
Make profile-wide model selection the single durable authority for gateway
sessions. Serialize global/session commits, verify the profile config write,
and clear the redundant durable and in-memory session override only after the
global write succeeds. Preserve truthful session fallback behavior when either
persistence step fails.

Keep gateway.authz_mixin as the public compatibility seam, move its existing
implementation behind the bounded model-authority mixin, and cover typed and
picker paths, ordering, concurrency, failure settlement, restart rehydration,
and the gpt-5.6-sol-900k context distinction.

Fixes #100314
Related #99808
Related #97963
